twenty third commit

+++ src/database/IConstDatabase.java
@@ -8,9 +8,8 @@ public interface IConstDatabase {
 	String SQL_INSERT_USER = new String("INSERT INTO user(name, surname, birth_day, birth_month, birth_year,"
 			+ " sex, login, password) VALUES(?,?,?,?,?,?,?,?)");
 	String SQL_INSERT_IMAGE = new String("INSERT INTO images(user_id, image_path) VALUES(?,?)");
-	String SQL_SELECT_USER = new String("SELECT * FROM user ");
-	String SQL_SELECT_IMAGE = new String("SELECT * FROM images ");
-	String SQL_DELETE_FROM_USER = new String("DELETE FROM user ");
-	String SQL_DELETE_FROM_IMAGES = new String("DELETE FROM images "); 
+	String SQL_SELECT_USER = new String("SELECT * FROM user WHERE login=?");
+	String SQL_SELECT_IMAGE = new String("SELECT * FROM images WHERE user_id=?");
+	String SQL_DELETE_FROM_IMAGES = new String("DELETE FROM images WHERE user_id=? AND image_path=?");
 
 }
